created the form and confirmed that form params are passed to the post action method

+++ CS460/HW4/MyFirstMVCProject/MyFirstMVCProject/Controllers/HomeController.cs
@@ -121,7 +121,7 @@ namespace MyFirstMVCProject.Controllers
 
         //POST ~/Home/Page3
         [HttpPost]
-        public ActionResult Page3(decimal amount, double rate, int term)
+        public ActionResult Page3(decimal amount, decimal rate, int term)
         {
             return View();
         }
